:truck: Move fixtures to main init for reuse
